+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Invalid revision when cancelling edit mode without changes
 - Set prompt library defaults on first language model added
 - Unable to cancel changes with active editor closed
+- Listener already registered warning when using Voqal VAD
 
 ## [2024.10.1] - 2024-09-05